+++ pycln/utils/scan.py
@@ -35,6 +35,7 @@ SUBSCRIPT_TYPE_VARIABLE = frozenset(
         "AsyncIterable",
         "AsyncIterator",
         "Awaitable",
+        "ByteString",
         "Callable",
         "ChainMap",
         "ClassVar",
@@ -54,7 +55,6 @@ SUBSCRIPT_TYPE_VARIABLE = frozenset(
         "Iterator",
         "KeysView",
         "List",
-        "Literal",  # Python >=3.7.
         "Mapping",
         "MappingView",
         "Match",
@@ -62,7 +62,6 @@ SUBSCRIPT_TYPE_VARIABLE = frozenset(
         "MutableSequence",
         "MutableSet",
         "Optional",
-        "OrderedDict",  # Python >=3.8.
         "Pattern",
         "Reversible",
         "Sequence",
@@ -72,6 +71,17 @@ SUBSCRIPT_TYPE_VARIABLE = frozenset(
         "Type",
         "Union",
         "ValuesView",
+        # Python >=3.7:
+        "Literal",
+        # Python >=3.8:
+        "OrderedDict",
+        # Python >=3.9:
+        "tuple",
+        "list",
+        "dict",
+        "set",
+        "frozenset",
+        "type",
     }
 )
 
@@ -184,10 +194,7 @@ class SourceAnalyzer(ast.NodeVisitor):
             getattr(func, "attr", "") == "cast"
             and getattr(func.value, "id", "") == "typing"  # type: ignore
         ):
-            type_ = node.args[0]
-            value = getattr(type_, "value", "") or getattr(type_, "s", "")
-            if value:
-                self._source_stats.name_.add(value)
+            self._parse_string(node.args[0])  # type: ignore
 
     @recursive
     def visit_Subscript(self, node: ast.Subscript) -> None:
@@ -199,13 +206,11 @@ class SourceAnalyzer(ast.NodeVisitor):
         #: Issue: https://github.com/hadialqattan/pycln/issues/32
         value = getattr(node, "value", "")
         if getattr(value, "id", "") in SUBSCRIPT_TYPE_VARIABLE or (
-            hasattr(value, "value") and getattr(value.value, "id") == "typing"
+            hasattr(value, "value") and getattr(value.value, "id", "") == "typing"
         ):
             s_val = node.slice.value  # type: ignore
             for elt in getattr(s_val, "elts", ()) or (s_val,):
-                value = getattr(elt, "value", "") or getattr(elt, "s", "")
-                if value:
-                    self._source_stats.name_.add(value)
+                self._parse_string(elt)
 
     @recursive
     def visit_Try(self, node: ast.Try):
@@ -315,14 +320,7 @@ class SourceAnalyzer(ast.NodeVisitor):
             annotation = node.annotation
         else:
             annotation = node.returns
-        if isinstance(annotation, (ast.Constant, ast.Str)):
-            if hasattr(annotation, "value"):
-                value = annotation.value
-            else:
-                value = annotation.s
-            if value:
-                tree = parse_ast(value, mode="eval")
-                self._add_name_attr(tree)
+        self._parse_string(annotation)  # type: ignore
 
     def _visit_type_comment(
         self, node: Union[ast.Assign, ast.arg, FunctionDefT]
@@ -342,6 +340,14 @@ class SourceAnalyzer(ast.NodeVisitor):
                 mode = "func_type"
             tree = parse_ast(type_comment, mode=mode)
             self._add_name_attr(tree)
+
+    def _parse_string(self, node: Union[ast.Constant, ast.Str]) -> None:
+        # Parse string names/attrs.
+        if isinstance(node, (ast.Constant, ast.Str)):
+            val = getattr(node, "value", "") or getattr(node, "s", "")
+            if val and isinstance(val, str):
+                tree = parse_ast(val, mode="eval")
+                self._add_name_attr(tree)
 
     def _add_name_attr(self, tree: ast.AST):
         # Add any `ast.Name` or `ast.Attribute`